refactor(LC002): split premature materialization helpers
- separate origin resolution and method-rule analysis
- keep analyzer behavior unchanged while reducing local complexity
