+++ src/LinqContraband/Analyzers/MaterializationAndProjection/LC002_PrematureMaterialization/PrematureMaterializationAnalyzer.cs
@@ -12,7 +12,7 @@ namespace LinqContraband.Analyzers.LC002_PrematureMaterialization;
 /// Analyzes query continuations that happen only after an IQueryable has already been materialized. Diagnostic ID: LC002
 /// </summary>
 [DiagnosticAnalyzer(LanguageNames.CSharp)]
-public sealed class PrematureMaterializationAnalyzer : DiagnosticAnalyzer
+public sealed partial class PrematureMaterializationAnalyzer : DiagnosticAnalyzer
 {
     public const string DiagnosticId = "LC002";
     private const string Category = "Performance";
@@ -146,45 +146,6 @@ public sealed class PrematureMaterializationAnalyzer : DiagnosticAnalyzer
             Diagnostic.Create(Rule, invocation.Syntax.GetLocation(), properties, invocation.TargetMethod.Name));
     }
 
-    private static bool TryCreateRedundantDiagnostic(
-        OperationAnalysisContext context,
-        IInvocationOperation invocation,
-        IOperation receiver,
-        out Diagnostic diagnostic)
-    {
-        diagnostic = null!;
-        if (!IsMaterializingMethod(invocation.TargetMethod)) return false;
-
-        if (!TryResolveMaterializationOrigin(
-                receiver,
-                invocation.Syntax.SpanStart,
-                context.Operation.FindOwningExecutableRoot(),
-                new HashSet<ILocalSymbol>(SymbolEqualityComparer.Default),
-                out var previousMaterialization))
-        {
-            return false;
-        }
-
-        var properties = CreateProperties(
-            RedundantDiagnosticKind,
-            previousMaterialization.OriginKind,
-            invocation.TargetMethod.Name,
-            previousMaterialization.MaterializerName);
-
-        if (CanOfferRemoveRedundantMaterializationFix(invocation.TargetMethod.Name, previousMaterialization))
-        {
-            properties = properties.SetItem(FixKindKey, RemoveRedundantMaterializationFixKind);
-        }
-
-        diagnostic = Diagnostic.Create(
-            RedundantRule,
-            invocation.Syntax.GetLocation(),
-            properties,
-            invocation.TargetMethod.Name,
-            previousMaterialization.MaterializerName);
-        return true;
-    }
-
     private static ImmutableDictionary<string, string?> CreateProperties(
         string diagnosticKind,
         string originKind,
@@ -198,256 +159,6 @@ public sealed class PrematureMaterializationAnalyzer : DiagnosticAnalyzer
             .Add(MaterializerKey, materializer);
     }
 
-    private static bool TryResolveMaterializationOrigin(
-        IOperation operation,
-        int position,
-        IOperation? executableRoot,
-        HashSet<ILocalSymbol> visitedLocals,
-        out MaterializationOrigin origin)
-    {
-        origin = default;
-        var unwrapped = operation.UnwrapConversions();
-
-        if (unwrapped is ILocalReferenceOperation localReference)
-        {
-            if (executableRoot == null) return false;
-
-            if (!TryResolveSingleAssignedValue(executableRoot, localReference.Local, position, visitedLocals, out var assignedValue))
-                return false;
-
-            if (TryResolveMaterializationOrigin(
-                    assignedValue,
-                    position,
-                    executableRoot,
-                    visitedLocals,
-                    out var localOrigin))
-            {
-                origin = new MaterializationOrigin(LocalOriginKind, localOrigin.MaterializerName);
-                return true;
-            }
-
-            return false;
-        }
-
-        if (unwrapped is IInvocationOperation materializerInvocation &&
-            IsMaterializingMethod(materializerInvocation.TargetMethod) &&
-            TryResolveQueryableOrMaterializedSource(
-                materializerInvocation.GetInvocationReceiver(),
-                position,
-                executableRoot,
-                visitedLocals))
-        {
-            origin = new MaterializationOrigin(InlineInvocationOriginKind, materializerInvocation.TargetMethod.Name);
-            return true;
-        }
-
-        if (unwrapped is IObjectCreationOperation objectCreation &&
-            objectCreation.Constructor != null &&
-            IsMaterializingConstructor(objectCreation.Constructor) &&
-            objectCreation.Arguments.Length > 0 &&
-            TryResolveQueryableOrMaterializedSource(
-                objectCreation.Arguments[0].Value,
-                position,
-                executableRoot,
-                visitedLocals))
-        {
-            origin = new MaterializationOrigin(ConstructorOriginKind, objectCreation.Constructor.ContainingType.Name);
-            return true;
-        }
-
-        return false;
-    }
-
-    private static bool TryResolveQueryableOrMaterializedSource(
-        IOperation? operation,
-        int position,
-        IOperation? executableRoot,
-        HashSet<ILocalSymbol> visitedLocals)
-    {
-        if (operation == null) return false;
-
-        var unwrapped = operation.UnwrapConversions();
-
-        if (unwrapped.Type?.IsIQueryable() == true) return true;
-
-        if (TryResolveMaterializationOrigin(unwrapped, position, executableRoot, visitedLocals, out _))
-        {
-            return true;
-        }
-
-        if (unwrapped is not ILocalReferenceOperation localReference || executableRoot == null)
-        {
-            return false;
-        }
-
-        return TryResolveSingleAssignedValue(executableRoot, localReference.Local, position, visitedLocals, out var assignedValue) &&
-               TryResolveQueryableOrMaterializedSource(assignedValue, position, executableRoot, visitedLocals);
-    }
-
-    private static bool TryResolveSingleAssignedValue(
-        IOperation executableRoot,
-        ILocalSymbol local,
-        int position,
-        HashSet<ILocalSymbol> visitedLocals,
-        out IOperation value)
-    {
-        value = null!;
-        if (!visitedLocals.Add(local)) return false;
-
-        IOperation? latestValue = null;
-        var latestPosition = -1;
-        var assignmentCount = 0;
-
-        foreach (var descendant in executableRoot.Descendants())
-        {
-            if (descendant.Syntax.SpanStart >= position) continue;
-
-            switch (descendant)
-            {
-                case IVariableDeclaratorOperation declarator when
-                    SymbolEqualityComparer.Default.Equals(declarator.Symbol, local) &&
-                    declarator.Initializer != null &&
-                    declarator.Syntax.SpanStart > latestPosition:
-                    latestValue = declarator.Initializer.Value;
-                    latestPosition = declarator.Syntax.SpanStart;
-                    assignmentCount++;
-                    break;
-
-                case ISimpleAssignmentOperation assignment when
-                    assignment.Target is ILocalReferenceOperation targetLocal &&
-                    SymbolEqualityComparer.Default.Equals(targetLocal.Local, local) &&
-                    assignment.Syntax.SpanStart > latestPosition:
-                    latestValue = assignment.Value;
-                    latestPosition = assignment.Syntax.SpanStart;
-                    assignmentCount++;
-                    break;
-            }
-        }
-
-        if (latestValue == null || assignmentCount != 1)
-        {
-            return false;
-        }
-
-        value = latestValue.UnwrapConversions();
-        return true;
-    }
-
-    private static bool IsApprovedContinuationMethod(IMethodSymbol method, out bool returnsSequence)
-    {
-        returnsSequence = false;
-
-        if (method.ContainingType.Name != "Enumerable" ||
-            method.ContainingNamespace?.ToString() != "System.Linq")
-        {
-            return false;
-        }
-
-        if (HasUnsupportedComparerOverload(method) || HasUnsupportedIndexAwareLambda(method))
-        {
-            return false;
-        }
-
-        if (SequenceContinuationMethods.Contains(method.Name))
-        {
-            returnsSequence = true;
-            return true;
-        }
-
-        if (TerminalContinuationMethods.Contains(method.Name))
-        {
-            return true;
-        }
-
-        return false;
-    }
-
-    private static bool HasUnsupportedComparerOverload(IMethodSymbol method)
-    {
-        if (method.Parameters.Length == 0) return false;
-
-        var lastParameterType = method.Parameters[method.Parameters.Length - 1].Type;
-        return lastParameterType.Name is "IEqualityComparer" or "IComparer" &&
-               lastParameterType.ContainingNamespace?.ToString() == "System.Collections.Generic";
-    }
-
-    private static bool HasUnsupportedIndexAwareLambda(IMethodSymbol method)
-    {
-        if (method.Parameters.Length == 0) return false;
-
-        var lastParameterType = method.Parameters[method.Parameters.Length - 1].Type as INamedTypeSymbol;
-        var invokeMethod = lastParameterType?.DelegateInvokeMethod;
-        return invokeMethod != null && invokeMethod.Parameters.Length > 1;
-    }
-
-    private static bool CanOfferMoveBeforeMaterializationFix(IMethodSymbol continuationMethod, MaterializationOrigin origin)
-    {
-        if (origin.OriginKind != InlineInvocationOriginKind) return false;
-        if (!IsApprovedContinuationMethod(continuationMethod, out _)) return false;
-
-        return origin.MaterializerName is
-            "ToList" or
-            "ToArray" or
-            "AsEnumerable" or
-            "ToImmutableList" or
-            "ToImmutableArray";
-    }
-
-    private static bool CanOfferRemoveRedundantMaterializationFix(string currentMaterializer, MaterializationOrigin previousMaterialization)
-    {
-        if (previousMaterialization.OriginKind != InlineInvocationOriginKind) return false;
-
-        return currentMaterializer == "AsEnumerable" ||
-               previousMaterialization.MaterializerName == "AsEnumerable" ||
-               (IsDirectCollectionMaterializer(currentMaterializer) &&
-                IsDirectCollectionMaterializer(previousMaterialization.MaterializerName));
-    }
-
-    private static bool IsMaterializingMethod(IMethodSymbol method)
-    {
-        var ns = method.ContainingNamespace?.ToString();
-        if (ns is not ("System.Linq" or "Microsoft.EntityFrameworkCore" or "System.Collections.Immutable")) return false;
-
-        if (method.Name == "AsEnumerable") return true;
-
-        return method.Name == "ToList" ||
-               method.Name == "ToListAsync" ||
-               method.Name == "ToArray" ||
-               method.Name == "ToArrayAsync" ||
-               method.Name == "ToDictionary" ||
-               method.Name == "ToDictionaryAsync" ||
-               method.Name == "ToHashSet" ||
-               method.Name == "ToHashSetAsync" ||
-               method.Name == "ToLookup" ||
-               method.Name.StartsWith("ToImmutable", StringComparison.Ordinal);
-    }
-
-    private static bool IsDirectCollectionMaterializer(string methodName)
-    {
-        return methodName is
-            "ToList" or
-            "ToArray" or
-            "ToHashSet" or
-            "ToImmutableList" or
-            "ToImmutableArray" or
-            "ToImmutableHashSet";
-    }
-
-    private static bool IsMaterializingConstructor(IMethodSymbol constructor)
-    {
-        var type = constructor.ContainingType;
-        if (type.ContainingNamespace?.ToString() != "System.Collections.Generic") return false;
-
-        return type.Name == "List" ||
-               type.Name == "HashSet" ||
-               type.Name == "Dictionary" ||
-               type.Name == "SortedDictionary" ||
-               type.Name == "SortedList" ||
-               type.Name == "LinkedList" ||
-               type.Name == "Queue" ||
-               type.Name == "Stack";
-    }
-
     private readonly struct MaterializationOrigin
     {
         public MaterializationOrigin(string originKind, string materializerName)
@@ -457,7 +168,6 @@ public sealed class PrematureMaterializationAnalyzer : DiagnosticAnalyzer
         }
 
         public string OriginKind { get; }
-
         public string MaterializerName { get; }
     }
 }
